+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Introduce the `VerifyData` trait to determine if data wrapped in a `Chunk` message should be propagated or not.
+
 ## [0.3.0]
 
 ### Added